ADD void pointer, null pointer constant fix array->pointer conversion

+++ test/039.c
@@ -1,7 +1,7 @@
 int main() {
-  short x;
+  unsigned short x;
   x = 65535;
-  int xx;
+  short xx;
   xx = x;
   printf("%d\n", xx);
